% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{rate2prob}
\alias{rate2prob}
\title{Miscellaneous functions to support the \code{ibmcraftr} packare are here.}
\usage{
rate2prob(rates)
}
\arguments{
\item{rates}{A numeric scalar or vector to be transformed into rates.}
}
\value{
A numeric scalar or vector in terms of probabilities.
}
\description{
Miscellaneous functions to support the \code{ibmcraftr} packare are here.
}
\examples{
rate2prob(c(.1, .5))

}

