% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syn_pop.R
\name{syn_pop}
\alias{syn_pop}
\title{Create a synthetic population having several states.}
\usage{
syn_pop(states, shuffle = FALSE)
}
\arguments{
\item{states}{A numeric vector with each element representing the number of individuals in a particular state
its index corresponds to.}

\item{shuffle}{A logical value to enable shuffling of the individuals (rows) in the resulting matrix.}
}
\value{
A matrix of 0s, and 1s. The rows representing the individuals and the columns representing the states
    the individuals are in
}
\description{
Populate a matrix in which columns represent the states of the individuals and rows represent the individuals.
}
\examples{
syn_pop(c(3,2,1))
syn_pop(c(0,0,1,5), shuffle=TRUE)

}

